bugfix/code changes after the PR comments
